# backend/src/main/resources/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "email": "info@springonly.com",
      "name": "Equipo de Desarrollo de Spring Only"
    },
    "description": "API para información y gestión de créditos simples.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Credit API REST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://backend.springonly.com",
      "description": "Servidor Spring Only"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateLoanRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "currencyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanDisbursementDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "numberOfMonthlyPayments": {
            "type": "integer",
            "format": "int16"
          }
        }
      },
      "CreateLoanResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "type": "string"
          },
          "currencyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanDisbursementDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "numberOfMonthlyPayments": {
            "type": "integer",
            "format": "int16"
          },
          "loanStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
      "CreatePaymentRequest": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "principalAmount": {
            "type": "number"
          },
          "interestAmount": {
            "type": "number"
          }
        }
      },
      "CreatePaymentResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentNumber": {
            "type": "integer",
            "format": "int16"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "principalAmount": {
            "type": "number"
          },
          "interestAmount": {
            "type": "number"
          },
          "totalPaymentAmount": {
            "type": "number"
          },
          "paymentStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "GetPaymentByIdResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentNumber": {
            "type": "integer",
            "format": "int16"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "principalAmount": {
            "type": "number"
          },
          "interestAmount": {
            "type": "number"
          },
          "totalPaymentAmount": {
            "type": "number"
          },
          "paymentStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "UpdatePaymentRequest": {
        "type": "object",
        "properties": {
          "paymentStateId": {
            "type": "string"
          }
        }
      },
      "UpdatePaymentResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentNumber": {
            "type": "integer",
            "format": "int16"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "principalAmount": {
            "type": "number"
          },
          "interestAmount": {
            "type": "number"
          },
          "totalPaymentAmount": {
            "type": "number"
          },
          "paymentStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/loans": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Create a new loan",
        "description": "Creates a loan for a customer. Requires a CreateLoanRequest payload with the necessary loan data.",
        "operationId": "createLoan",
        "requestBody": {
          "description": "CreateLoanRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "Create loan example": {
                  "summary": "Request to create a loan",
                  "value": {
                    "customerId": "CUST12345",
                    "currencyId": "PEN",
                    "principalAmount": 10000.0,
                    "interestRate": 5.5,
                    "loanDisbursementDate": "2025-09-01",
                    "numberOfMonthlyPayments": 36
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateLoanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Loan created successfully",
            "content": {
              "application/json": {
                "examples": {
                  "CreateLoanResponse example": {
                    "summary": "Created loan representation",
                    "value": {
                      "loanId": 12345,
                      "customerId": "CUST12345",
                      "currencyId": "PEN",
                      "principalAmount": 10000.0,
                      "interestRate": 5.5,
                      "loanDisbursementDate": "2025-09-01",
                      "numberOfMonthlyPayments": 36,
                      "loanStateId": "EN EVALUACION",
                      "writtenAt": "2025-10-07T22:57:09-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/CreateLoanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "CreateLoan error": {
                    "summary": "Invalid input example",
                    "value": {
                      "message": "principalAmount must be greater than 0",
                      "code": "INVALID_REQUEST"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Loan Resource"
        ]
      }
    },
...
    "/api/loans/{loanId}/payments": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "List payments",
        "description": "List all payments associated to a given loan.",
        "operationId": "listPaymentsByLoanId",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payments list for loan",
            "content": {
              "application/json": {
                "examples": {
                  "Payments list": {
                    "summary": "A list of payments for a loan",
                    "value": [
                      {
                        "loanId": 12345,
                        "paymentNumber": 1,
                        "dueDate": "2025-10-15",
                        "principalAmount": 500.0,
                        "interestAmount": 5.0,
                        "totalPaymentAmount": 505.0,
                        "paymentStateId": "PAGADO",
                        "writtenAt": "2025-08-16T13:25:00-05:00"
                      },
                      {
                        "loanId": 12345,
                        "paymentNumber": 2,
                        "dueDate": "2025-11-15",
                        "principalAmount": 500.0,
                        "interestAmount": 4.5,
                        "totalPaymentAmount": 504.5,
                        "paymentStateId": "PENDIENTE",
                        "writtenAt": "2025-07-16T11:15:00-05:00"
                      }
                    ]
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetPaymentByIdResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No payments found for loan",
            "content": {
              "application/json": {
                "examples": {
                  "No payments found": {
                    "summary": "No payments for given loan",
                    "value": {
                      "message": "El préstamo no tiene pagos registrados",
                      "code": "PAY002"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Payment Resource"
        ]
      }
    },
    "/api/loans/{loanId}/payments/{paymentNumber}": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "summary": "Update payment",
        "description": "Update an existing payment for a loan identified by loanId and paymentNumber.",
        "operationId": "updatePayment",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "paymentNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int16"
            }
          }
        ],
        "requestBody": {
          "description": "UpdatePaymentRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "UpdatePayment request": {
                  "summary": "Request to update payment state",
                  "value": {
                    "paymentStateId": "PAGADO"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payment updated successfully",
            "content": {
              "application/json": {
                "examples": {
                  "UpdatePaymentResponse example": {
                    "summary": "Updated payment representation",
                    "value": {
                      "loanId": 12345,
                      "paymentNumber": 1,
                      "dueDate": "2025-10-15",
                      "principalAmount": 500.0,
                      "interestAmount": 5.0,
                      "totalPaymentAmount": 505.0,
                      "paymentStateId": "PAGADO",
                      "writtenAt": "2025-09-16T15:45:00-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/UpdatePaymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found",
            "content": {
              "application/json": {
                "examples": {
                  "UpdatePayment not found": {
                    "summary": "Payment to update was not found",
                    "value": {
                      "message": "El pago indicado no existe para este préstamo",
                      "code": "PAY404"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Payment Resource"
        ]
      },
      "get": {
        "summary": "Get payment by id",
        "description": "Retrieve a specific payment for a loan by loanId and paymentNumber.",
        "operationId": "getPaymentById",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "paymentNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int16"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment retrieved successfully",
            "content": {
              "application/json": {
                "examples": {
                  "GetPaymentByIdResponse example": {
                    "summary": "A payment retrieved by id",
                    "value": {
                      "loanId": 12345,
                      "paymentNumber": 1,
                      "dueDate": "2025-10-15",
                      "principalAmount": 500.0,
                      "interestAmount": 5.0,
                      "totalPaymentAmount": 505.0,
                      "paymentStateId": "PAGADO",
                      "writtenAt": "2025-09-16T15:45:00-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found",
            "content": {
              "application/json": {
                "examples": {
                  "GetPaymentById not found": {
                    "summary": "Requested payment does not exist",
                    "value": {
                      "message": "El pago solicitado no existe",
                      "code": "PAY003"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Payment Resource"
        ]
      },
      "post": {
        "summary": "Create payment",
        "description": "Create a payment for a specific loan and payment number.",
        "operationId": "createPayment",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "paymentNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int16"
            }
          }
        ],
        "requestBody": {
          "description": "CreatePaymentRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "CreatePayment request": {
                  "summary": "Request to create a payment",
                  "value": {
                    "dueDate": "2025-10-15",
                    "principalAmount": 500.0,
                    "interestAmount": 5.0
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Payment created successfully",
            "content": {
              "application/json": {
                "examples": {
                  "CreatePaymentResponse example": {
                    "summary": "Created payment representation",
                    "value": {
                      "loanId": 12345,
                      "paymentNumber": 1,
                      "dueDate": "2025-10-15",
                      "principalAmount": 500.0,
                      "interestAmount": 5.0,
                      "totalPaymentAmount": 505.0,
                      "paymentStateId": "PENDIENTE",
                      "writtenAt": "2025-09-16T15:45:00-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "CreatePayment error": {
                    "summary": "Invalid create payment input",
                    "value": {
                      "message": "Invalid input data",
                      "code": "PAY400"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Payment Resource"
        ]
      }
    },
...
  }
}